Se corrigió el tipo de dato de 'hours_of_use' a DECIMAL en el modelo de UserAppliance y se mejoró la consistencia en la indentación del modelo de MonthlyConsumption.

+++ backend/src/models/monthlyConsumption.model.js
@@ -1,31 +1,35 @@
-import { DataTypes } from 'sequelize';
-import { sequelize } from '../config/database.js'; // Ajusta la ruta si es diferente
+import { DataTypes } from "sequelize";
+import { sequelize } from "../config/database.js"; // Ajusta la ruta si es diferente
 
-export const MonthlyConsumptionModel = sequelize.define('MonthlyConsumption', {
+export const MonthlyConsumptionModel = sequelize.define(
+  "MonthlyConsumption",
+  {
     id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true,
+      type: DataTypes.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
     },
     user_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-            model: 'Users', // Ajusta si el nombre de la tabla es diferente
-            key: 'id',
-        },
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      references: {
+        model: "Users", // Ajusta si el nombre de la tabla es diferente
+        key: "id",
+      },
     },
     month: {
-        type: DataTypes.STRING, // Formato 'YYYY-MM'
-        allowNull: false,
+      type: DataTypes.STRING, // Formato 'YYYY-MM'
+      allowNull: false,
     },
     cost: {
-        type: DataTypes.FLOAT,
-        allowNull: false,
+      type: DataTypes.DECIMAL(10, 2),
+      allowNull: false,
     },
-}, {
-    tableName: 'monthly_consumptions',
+  },
+  {
+    tableName: "monthly_consumptions",
     timestamps: true,
-});
+  }
+);
 
 export default MonthlyConsumptionModel;
